Implement GRPC client for order-service

+++ common-libs/src/main/java/com/emercurius/commonlibs/dto/order/OrderItemResponseDTO.java
@@ -8,7 +8,7 @@ public record OrderItemResponseDTO(
 
         String productName,
 
-        Long quantity,
+        int quantity,
 
         Long orderId,
